Factor out commonalities in retry and repeat
